+++ Source/Core/CrPlatform.h
@@ -5,7 +5,7 @@ namespace cr
 	// Relevant platforms here. The platform and the graphics API are kept separate
 	namespace Platform
 	{
-		enum T : uint8_t
+		enum T : uint32_t
 		{
 			Windows,
 			Count
